+++ lib/presentation/onboarding/views/baby_registration_view.dart
@@ -7,7 +7,7 @@ import 'package:yuktoe/common/design_system/app_text_styles.dart';
 import 'package:yuktoe/constants/app_strings.dart';
 import 'package:yuktoe/constants/enum/gender.dart';
 import 'package:yuktoe/domain/models/baby_registration/onboarding_flow.dart';
-import 'package:yuktoe/onboarding/view_models/baby_registration_view_model.dart';
+import 'package:yuktoe/presentation/onboarding/view_models/baby_registration_view_model.dart';
 import 'package:yuktoe/routing/router.dart';
 
 class BabyRegistrationView extends StatelessWidget {
