+++ test/onboarding/view_models/baby_profile_setup_view_model_test.dart
@@ -7,7 +7,7 @@ import 'package:yuktoe/constants/enum/gender.dart';
 import 'package:yuktoe/constants/enum/relationship.dart';
 import 'package:yuktoe/data/repositories/baby_registration_repository/baby_registration_repository.dart';
 import 'package:yuktoe/domain/models/baby_registration/onboarding_flow.dart';
-import 'package:yuktoe/onboarding/view_models/baby_profile_setup_view_model.dart';
+import 'package:yuktoe/presentation/onboarding/view_models/baby_profile_setup_view_model.dart';
 
 import 'baby_profile_setup_view_model_test.mocks.dart';
